% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_numerique.R
\name{integrationboole}
\alias{integrationboole}
\title{Title Integration numerique par la methode de Boole}
\usage{
integrationboole(f, a, b, n)
}
\arguments{
\item{f}{function to integrate}

\item{a}{lower bound of integration}

\item{b}{upper bound of integration}

\item{n}{number of subdivisions}
}
\value{
The approximate value of the integral
}
\description{
Title Integration numerique par la methode de Boole
}
\examples{
f <- function(x) x^2 * sin(x)
resultat <- integrationboole(f, 0, pi, 16)
}
