% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_numerique.R
\name{integrationsimpson}
\alias{integrationsimpson}
\title{Title Numerical Integration using Simpsons Rule}
\usage{
integrationsimpson(f, a, b, n)
}
\arguments{
\item{f}{function to integrate}

\item{a}{lower bound of integration}

\item{b}{upper bound of integration}

\item{n}{number of subdivisions (must be even)}
}
\value{
The approximate value of the integral
}
\description{
Title Numerical Integration using Simpsons Rule
}
\examples{
f <- function(x) sin(x)
integrationsimpson(f,0,pi,32)
}
