% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_numerique.R
\name{integrationtrapeze}
\alias{integrationtrapeze}
\title{Title Numerical Integration using the Trapezoidal Rule}
\usage{
integrationtrapeze(f, a, b, n)
}
\arguments{
\item{f}{function to integrate}

\item{a}{lower bound of integration}

\item{b}{upper bound of integration}

\item{n}{number of subdivisions}
}
\value{
value of the integral
}
\description{
Title Numerical Integration using the Trapezoidal Rule
}
\examples{
f <- function(x) 3*(x^3)
integrationtrapeze(f,0,1,10)
}
